Completed implementation of linear solve
This commit contains the complete version of a linear solver that
works at compile time using LU factorization with partial
pivoting. There is a non-trivial test case with a matrix that does
require pivoting for factorization to succeed. The main entry
point is `linear_solve(A, b)` to return `x` satisfying `Ax = b`.