Fix operator== for Terms
The old operator== on `Term`s only compared the powers;
while this didn't manifest in any bugs it is certainly
unexpected behavior. This commit fixes it. It also contains
changes to make multinomials work with negative powers as
well as positive ones. These changes are *not* complete yet.
I need to add tests for multinomial evaluation with negative powers,
and fix the behavior to be correct.