Impl. derivative DOF's
This commit adds a `PartialAt` struct and helper function
to create it to namespace `Elements`. This struct acts as a
function object to pass as a constraint to
`derive_shape_functions`
so that the values of partial derivatives at points may be used as
degrees of freedom to define an element. A test case is added,
deriving the shape functions for a 1-dimensional element
with cubic shape functions where the degrees of freedom
are the values of both the function and its derivative at the endpoints.
This element is useful in practice because the weak solution using this
function space is a subset of C1.