Refactor Multinomials -> Metanomials
This commit has large changes in the (now) Metanomial interface.
As preparation to introduce a new `Functions` namespace that will
at first encompass polynomials, but provide rules for calculus with
sums, products, compositions of functions, I changed the namespace
name (and class and function names) for Multinomials to Metanomials.
This distinguishes them from Polynomials, which will be runtime objects.

This commit also refactors the `partial` function, which was a member of
the Metanomials namespace, to be a free function `utils.hpp`. The free
function if called as `partial<I>(x)` just returns `x.partial<I>()`, so
when called it expects `x` to implement that interface. Metanomial
and Term classes implement the interface now.