Fixes for gcc
In checking assembly in a test file, I found that the two small changes
in this commit were necessary for gcc to accept the code. I think that
the fact that these changes are
required is a bug in gcc but I do not have time to figure it out.