Begin impl. of BilinearQuadTransform
This is the general form of the transformation from a reference square
(the domain [-1, 1] x [-1, 1]) to a convex quadrilateral. In this commit
I have the form of the transformation, with tests of the constructor
interface and of the application of the transform. Next is to add tests
(and implement) the Jacobian elements and determinant.