Add specialized classes for form_matrix return
Before this commit I planned to return closures from form_matrix.
I discovered in the symmetric case that I couldn't get clang to do this
without an extra 8 bytes in the closure wasted on capturing the matrix
size, which isn't needed. I created a symmetric matrix struct just to
return from form_matrix and avoid this issue, and I figured that the
asymmetric case should have a corresponding type to return.