Remove uses of std::get
In order to write generic code, my interfaces should only require a get
function template found via ADL for types passed to them. std::get is
not a customization point so overloads can't be added for custom types;
by avoiding calling std::get and instead using an unqualified name `get`
it is possible for users to define a `get` function in the same
namespace as their custom type and have it work.