Add the basics of TriangleTransform
This commit is a record of where I got to in my first 20 minute session
working on the triangle transform. It has the functionality to construct
the transformation and map points using it, with tests. On Monday I
need to add the Jacobian functionality, etc. to make it useful but for
now I am going to bed.